+++ Mods ChoGGi/All Sponsor Buildings/items.lua
@@ -7,7 +7,6 @@ if not g_AvailableDlc.gagarin then
 		PlaceObj("ModItemOptionToggle", {
 			"name", "NODLC",
 			"DisplayName", T(0000, "Space Race DLC not installed!"),
-			"DefaultValue", true,
 		}),
 	}
 end
@@ -32,12 +31,36 @@ local techs = {
 }
 local TechDef = TechDef
 
+--~ 'AdvancedStirlingGenerator' = true
+--~ 'AutomaticMetalsExtractor' = true
+--~ 'ConcretePlant' = true
+--~ 'CorporateOffice' = true
+--~ 'GameDeveloper' = true
+--~ 'JumperShuttleHub' = true
+--~ 'LowGLab' = true
+--~ 'MegaMall' = true
+--~ 'MetalsRefinery' = true
+--~ 'RareMetalsRefinery' = true
+--~ 'RCConstructorBuilding' = true
+--~ 'RCDrillerBuilding' = true
+--~ 'RCHarvesterBuilding' = true
+--~ 'RCSensorBuilding' = true
+--~ 'RCSolarBuilding' = true
+--~ 'ShuttleHub' = true
+--~ 'SolarArray' = true
+--~ 'TaiChiGarden' = true
+--~ 'Temple' = true
+
 local BuildingTemplates = BuildingTemplates
 for id, bld in pairs(BuildingTemplates) do
 	for i = 1, 3 do
-		if bld["sponsor_status" .. i] ~= false
+--~ 		if bld["sponsor_status" .. i] ~= false
+--~ 			-- Shuttle Hub has "disabled" to block it when Jumper (paradox spons) is enabled.
+--~ 			and bld["sponsor_status" .. i] ~= "disabled"
+--~ 		then
+		if bld["sponsor_name" .. i] ~= ""
 			-- Shuttle Hub has "disabled" to block it when Jumper (paradox spons) is enabled.
-			and bld["sponsor_status" .. i] ~= "disabled"
+			and bld.id ~= "ShuttleHub"
 		then
 
 			-- Get an image to use for desc
@@ -65,7 +88,8 @@ for id, bld in pairs(BuildingTemplates) do
 				mod_options[c] = PlaceObj("ModItemOptionToggle", {
 					"name", "ChoGGi_Tech_" .. id,
 					"DisplayName", table.concat(T(bld.display_name) .. " " .. T(3734--[[Tech]])),
-					"Help", table.concat(T(0000, "Lock behind tech unlock.")
+					"Help", table.concat(
+						T(0000, "Lock behind tech unlock.")
 						.. "\n\n" .. def.display_name
 						.. "\n\n<image " .. def.icon .. ">"
 					),
@@ -84,4 +108,5 @@ table.sort(mod_options, function(a, b)
 	return CmpLower(_InternalTranslate(a.DisplayName), _InternalTranslate(b.DisplayName))
 end)
 
+--~ ex{"mod_options",mod_options}
 return mod_options