+++ Mods ChoGGi/Blank Sponsor/metadata.lua
@@ -1,0 +1,27 @@
+return PlaceObj("ModDef", {
+	"title", "Blank Sponsor",
+	"id", "ChoGGi_BlankSponsor",
+	"steam_id", "3394519699",
+	"pops_any_uuid", "4ce65347-c6c1-473e-8790-2d0e6486d9f0",
+	"lua_revision", 1007000, -- Picard
+	"version", 2,
+	"version_major", 0,
+	"version_minor", 2,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagCommanderProfiles", true,
+	"TagMissionSponsors", true,
+	"description", [[
+Sponsor/Commander profile with no extras added (buildings/etc).
+
+Because it's no fun just doing that, this also does random mission goals for a bit of variety.
+There's also a mod option to have random goals for any sponsor.
+
+
+Requested by Longjumping_Fold_815.
+]],
+})