+++ Mods ChoGGi/Project Phoenix No Renegades/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2463513394",
 	"pops_any_uuid", "6967fd74-c9b3-4c47-95e9-b38a494572ba",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {