+++ Mods ChoGGi/Real Time Clock/Code/Script.lua
@@ -44,7 +44,7 @@ local function SetClock()
 	end
 end
 
-
+-- Define func vars now to be able to use them above where they're defined.
 local AddTime
 local RemoveTime
 
@@ -153,8 +153,10 @@ end
 local ChoOrig_OpenDialog = OpenDialog
 function OpenDialog(dlg_str, ...)
 	local dlg = ChoOrig_OpenDialog(dlg_str, ...)
-	if mod_ShowClock and dlg_str == "Infobar" then
-		CreateRealTimeThread(AddTime, dlg)
+	if mod_ShowClock and dlg_str == "Infobar"
+		and not IsValidThread(dlg.ChoGGi_RealTimeClock)
+	then
+		dlg.ChoGGi_RealTimeClock = CreateRealTimeThread(AddTime, dlg)
 	end
 	return dlg
 end